requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
django-bmemcached==0.3.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
protobuf==3.19.1
psycopg2==2.9.2
pyparsing==3.0.6
python-binary-memcached==0.30.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uhashring==2.1
uritemplate==4.1.1
urllib3==1.26.7
